interview/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
bcrypt>=4.0
httpx>=0.26.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.0.0
python-multipart>=0.0.6
email-validator>=2.0
python-docx>=1.0.0
pdfplumber>=0.10.0
pypdf>=4.0.0
